特性新增：新增中断通知图像接受以及稳定PCIE驱动

+++ source/base/image_buffer.h
@@ -13,7 +13,7 @@ typedef struct {
     uint8_t  vc_id;
     uint16_t width;
     uint16_t height;
-    float    fps;
+    int    fps;
     uint32_t rsvd;
 } image_info_t;
